+++ src/locale/index.ts
@@ -31,6 +31,7 @@ export default {
   },
   image: {
     broken: 'Broken image',
+    preview: 'Preview image',
     replace: 'Replace',
   },
   link: {