+++ CHANGELOG.md
@@ -2,6 +2,110 @@
 
 # Changelog
 
+## [Version&nbsp;4.26.0](https://github.com/lobehub/lobe-editor/compare/v4.25.1...v4.26.0)
+<sup>Released on **2026-09-02**</sup>
+
+
+#### ✨ Features
+
+- **image**: Add preview action.
+
+
+<br/>
+
+
+
+<details>
+<summary><kbd>Improvements and Fixes</kbd></summary>
+
+
+
+#### What's improved
+
+* **image**: Add preview action, closes [#202](https://github.com/lobehub/lobe-editor/issues/202) ([3152bdc](https://github.com/lobehub/lobe-editor/commit/3152bdc))
+
+</details>
+
+
+<div align="right">
+
+[![](https://img.shields.io/badge/-BACK_TO_TOP-151515?style=flat-square)](#readme-top)
+
+</div>
+
+### [Version&nbsp;4.25.1](https://github.com/lobehub/lobe-editor/compare/v4.25.0...v4.25.1)
+<sup>Released on **2026-09-02**</sup>
+
+
+#### 🐛 Bug Fixes
+
+- **misc**: Release editor resources on root detach.
+
+
+<br/>
+
+
+
+<details>
+<summary><kbd>Improvements and Fixes</kbd></summary>
+
+
+
+#### What's fixed
+
+* **misc**: Release editor resources on root detach, closes [#205](https://github.com/lobehub/lobe-editor/issues/205) ([1423ee6](https://github.com/lobehub/lobe-editor/commit/1423ee6))
+
+</details>
+
+
+<div align="right">
+
+[![](https://img.shields.io/badge/-BACK_TO_TOP-151515?style=flat-square)](#readme-top)
+
+</div>
+
+## [Version&nbsp;4.25.0](https://github.com/lobehub/lobe-editor/compare/v4.24.1...v4.25.0)
+<sup>Released on **2026-09-02**</sup>
+
+
+#### ✨ Features
+
+- **misc**: Integrate collaboration and rich document capabilities.
+
+
+#### 🐛 Bug Fixes
+
+- **misc**: Self-provide motion context in LexicalRenderer.
+
+
+<br/>
+
+
+
+<details>
+<summary><kbd>Improvements and Fixes</kbd></summary>
+
+
+
+#### What's improved
+
+* **misc**: Integrate collaboration and rich document capabilities, closes [#198](https://github.com/lobehub/lobe-editor/issues/198) ([6a3f836](https://github.com/lobehub/lobe-editor/commit/6a3f836))
+
+
+
+#### What's fixed
+
+* **misc**: Self-provide motion context in LexicalRenderer, closes [#206](https://github.com/lobehub/lobe-editor/issues/206) ([04c3c0e](https://github.com/lobehub/lobe-editor/commit/04c3c0e))
+
+</details>
+
+
+<div align="right">
+
+[![](https://img.shields.io/badge/-BACK_TO_TOP-151515?style=flat-square)](#readme-top)
+
+</div>
+
 ### [Version&nbsp;4.24.1](https://github.com/lobehub/lobe-editor/compare/v4.24.0...v4.24.1)
 <sup>Released on **2026-08-19**</sup>